pyproject.toml
[tool.poetry]
name = "autodoc"
version = "0.1.0"
description = ""
authors = ["Vitaly Pechkunov <witalink.viphaus@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
selene = "2.0.0rc8"
python-dotenv = "1.0.1"
pytest = "8.0.1"
allure-pytest = "2.13.5"
requests = "2.31.0"
jsonschema = "4.21.1"
appium-python-client = "4.0.0"
curlify = "^2.2.1"

[tool.pytest.ini_options]
addopts = [
    "--alluredir", "allure-results",
    "--clean-alluredir",
    "-s",
    "-v",
]

testpaths = [
    "tests/api",
    "tests/mobile",
    "tests/ui",
    "tests/ui_api",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"